requirements.txt
opencv-python-headless==4.4.0.0
streamlit==0.73.1
tensorflow==2.4.0
numpy==1.18.3
Pillow==8.0.1
